Move to SFC, cleanup typo in routing
